- sha1_computeHashOfFile() now returns number of processed byted. - Fixed signed/unsigned warning.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@2359 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/scriptingclient.c
@@ -628,7 +628,7 @@ int main(int argc, char** argv)
       printf("Environment Name    = ");
       if(EnvironmentName) {
          printf("%s (", EnvironmentName);
-         for(i = 0; i < sizeof(EnvironmentHash); i++) {
+         for(i = 0; i < (int)sizeof(EnvironmentHash); i++) {
             printf("%02x", (unsigned int)EnvironmentHash[i]);
          }
          puts(")");
